+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.6.0-dev
+  **BUG FIXES**
+  - [#140](https://github.com/Datatamer/unify-client-python/issues/140) Dataset `itergeofeatures` now supports data with geo attribute NULL
 
 ## 0.5.0
   **NEW FEATURES**